changing comparison code to a different folder
