+++ simulations/triplet_packing_simulation/triplet_packing_cpp_vs_python/cpp_triplet_packing_files/solver_ls_wrapper.h
