+++ simulations/triplet_packing_simulation/triplet_packing_cpp_vs_python/results/backtracking_py_and_cpp_comparison.png
