+++ simulations/triplet_packing_simulation/triplet_packing_cpp_vs_python/results/comparison_py_triplet_vs_cpp_triplet_algo.csv
