+++ simulations/triplet_packing_simulation/triplet_packing_vs_bin_completion/results/bin_completion_vs_backtrack_runtime_comparison.png
