+++ simulations/triplet_packing_simulation/triplet_packing_vs_bin_completion/results/comparison_bin_completion_vs_triplet_packing.csv
