+++ simulations/triplet_packing_simulation/triplet_packing_vs_bin_completion/results/comparison_triplet_vs_local.csv
