+++ simulations/triplet_packing_simulation/triplet_packing_vs_bin_completion/results/triplet_local_vs_backtrack_runtime_comparison.png
